requirements.txt
asgiref==3.6.0
Django==4.2.1
django-admin-interface==0.26.0
django-ckeditor==6.5.1
django-colorfield==0.8.0
django-js-asset==2.0.0
django-render-partial==0.4
Pillow==9.5.0
psycopg2==2.9.6
python-slugify==8.0.1
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
